internal/plots: move new plots in plots.go
